requirements.txt
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==2.2.3
django-filter==2.3.0
django-rest-knox==4.1.0
djangorestframework==3.11.0
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
pycparser==2.20
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
urllib3==1.25.9
zipp==1.2.0
